+++ R/block.R
@@ -249,7 +249,9 @@ block_exec = function(options) {
   output = knit_hooks$get('chunk')(output, options)
 
   if (options$cache > 0) {
-    obj.new = setdiff(ls(globalenv(), all.names = TRUE), obj.before)
+    # if cache.vars has been specifically provided, only cache these vars and no
+    # need to look for objects in globalenv()
+    obj.new = if (is.null(options$cache.vars)) setdiff(ls(globalenv(), all.names = TRUE), obj.before)
     copy_env(globalenv(), env, obj.new)
     objs = if (isFALSE(ev) || length(code) == 0) character(0) else
       options$cache.vars %n% codetools::findLocalsList(parse_only(code))