+++ R/utils.R
@@ -708,7 +708,13 @@ current_input = function(dir = FALSE) {
 default_handlers = evaluate:::default_output_handler
 # change the value handler in evaluate default handlers
 knit_handlers = function(fun, options) {
-  if (!is.function(fun)) fun = knit_print
+  if (!is.function(fun)) fun = function(x, ...) {
+    res = knit_print(x, ...)
+    # indicate the htmlwidget result with a special class so we can attach
+    # the figure caption to it later in wrap.knit_asis
+    if (inherits(x, 'htmlwidget')) class(res) = c(class(res), 'knit_asis_htmlwidget')
+    res
+  }
   if (length(formals(fun)) < 2)
     stop("the chunk option 'render' must be a function of the form ",
          "function(x, options) or function(x, ...)")